Address windows -D warnings
In #7734 we're making `-D warnings` common across all relevant CI
jobs. This commit resolves warnings issued by our windows build. It's
particularly noisy compared to Unix, on account of Windows is not
represented as a dev platform in the team.
