+++ lib/shared/src/conversion.rs
@@ -6,6 +6,9 @@ use std::collections::{HashMap, HashSet};
 use std::fmt::Debug;
 use std::num::{ParseFloatError, ParseIntError};
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Debug, Snafu)]
 pub enum ConversionError {
     #[snafu(display("Unknown conversion name {:?}", name))]
@@ -240,168 +243,4 @@ fn parse_timestamp(tz: TimeZone, s: &str) -> Result<DateTime<Utc>, Error> {
         }
     }
     Err(Error::AutoTimestampParseError { s: s.into() })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::parse_bool;
-    #[cfg(unix)]
-    use super::parse_timestamp;
-    use super::Bytes;
-    #[cfg(unix)]
-    use super::{Conversion, Error};
-    use crate::datetime::TimeZone;
-    use chrono::{DateTime, NaiveDateTime, TimeZone as _, Utc};
-    use chrono_tz::{Australia, Tz};
-
-    #[cfg(unix)]
-    const TIMEZONE_NAME: &str = "Australia/Brisbane";
-    const TIMEZONE: Tz = Australia::Brisbane;
-
-    #[derive(PartialEq, Debug, Clone)]
-    enum StubValue {
-        Bytes(Bytes),
-        Timestamp(DateTime<Utc>),
-        Float(f64),
-        Integer(i64),
-        Boolean(bool),
-    }
-
-    impl From<Bytes> for StubValue {
-        fn from(v: Bytes) -> Self {
-            StubValue::Bytes(v)
-        }
-    }
-
-    impl From<DateTime<Utc>> for StubValue {
-        fn from(v: DateTime<Utc>) -> Self {
-            StubValue::Timestamp(v)
-        }
-    }
-
-    impl From<f64> for StubValue {
-        fn from(v: f64) -> Self {
-            StubValue::Float(v)
-        }
-    }
-
-    impl From<i64> for StubValue {
-        fn from(v: i64) -> Self {
-            StubValue::Integer(v)
-        }
-    }
-
-    impl From<bool> for StubValue {
-        fn from(v: bool) -> Self {
-            StubValue::Boolean(v)
-        }
-    }
-
-    #[cfg(unix)]
-    fn dateref() -> DateTime<Utc> {
-        Utc.from_utc_datetime(&NaiveDateTime::from_timestamp(981173106, 0))
-    }
-
-    #[cfg(unix)]
-    fn convert<T>(fmt: &str, value: &'static str) -> Result<T, Error>
-    where
-        T: From<Bytes> + From<i64> + From<f64> + From<bool> + From<DateTime<Utc>>,
-    {
-        std::env::set_var("TZ", TIMEZONE_NAME);
-        Conversion::parse(fmt, TimeZone::Local)
-            .unwrap_or_else(|_| panic!("Invalid conversion {:?}", fmt))
-            .convert(value.into())
-    }
-
-    #[cfg(unix)] // https://github.com/timberio/vector/issues/1201
-    #[test]
-    fn timestamp_conversion() {
-        assert_eq!(
-            convert::<StubValue>("timestamp", "02/03/2001:14:05:06"),
-            Ok(dateref().into())
-        );
-    }
-
-    #[cfg(unix)] // see https://github.com/timberio/vector/issues/1201
-    #[test]
-    fn timestamp_param_conversion() {
-        assert_eq!(
-            convert::<StubValue>("timestamp|%Y-%m-%d %H:%M:%S", "2001-02-03 14:05:06"),
-            Ok(dateref().into())
-        );
-    }
-
-    #[cfg(unix)] // see https://github.com/timberio/vector/issues/1201
-    #[test]
-    fn parse_timestamp_auto_tz_env() {
-        std::env::set_var("TZ", TIMEZONE_NAME);
-        let good = Ok(dateref());
-        let tz = TimeZone::Local;
-        assert_eq!(parse_timestamp(tz, "2001-02-03 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "02/03/2001:14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-03T14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-03T04:05:06Z"), good);
-        assert_eq!(parse_timestamp(tz, "Sat, 3 Feb 2001 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "Sat Feb 3 14:05:06 2001"), good);
-        assert_eq!(parse_timestamp(tz, "3-Feb-2001 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-02T22:05:06-06:00"), good);
-        assert_eq!(parse_timestamp(tz, "Sat, 03 Feb 2001 07:05:06 +0300"), good);
-        assert_eq!(parse_timestamp(tz, "03/Feb/2001:02:05:06 -0200"), good);
-    }
-
-    #[cfg(unix)] // see https://github.com/timberio/vector/issues/1201
-    #[test]
-    fn parse_timestamp_auto() {
-        let good = Ok(dateref());
-        let tz = TimeZone::Named(TIMEZONE);
-        assert_eq!(parse_timestamp(tz, "2001-02-03 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "02/03/2001:14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-03T14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-03T04:05:06Z"), good);
-        assert_eq!(parse_timestamp(tz, "Sat, 3 Feb 2001 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "Sat Feb 3 14:05:06 2001"), good);
-        assert_eq!(parse_timestamp(tz, "3-Feb-2001 14:05:06"), good);
-        assert_eq!(parse_timestamp(tz, "2001-02-02T22:05:06-06:00"), good);
-        assert_eq!(parse_timestamp(tz, "Sat, 03 Feb 2001 07:05:06 +0300"), good);
-    }
-
-    // These should perhaps each go into an individual test function to be
-    // able to determine what part failed, but that would end up really
-    // spamming the test logs.
-
-    #[test]
-    fn parse_bool_true() {
-        assert_eq!(parse_bool("true"), Ok(true));
-        assert_eq!(parse_bool("True"), Ok(true));
-        assert_eq!(parse_bool("t"), Ok(true));
-        assert_eq!(parse_bool("T"), Ok(true));
-        assert_eq!(parse_bool("yes"), Ok(true));
-        assert_eq!(parse_bool("YES"), Ok(true));
-        assert_eq!(parse_bool("y"), Ok(true));
-        assert_eq!(parse_bool("Y"), Ok(true));
-        assert_eq!(parse_bool("1"), Ok(true));
-        assert_eq!(parse_bool("23456"), Ok(true));
-        assert_eq!(parse_bool("-8"), Ok(true));
-    }
-
-    #[test]
-    fn parse_bool_false() {
-        assert_eq!(parse_bool("false"), Ok(false));
-        assert_eq!(parse_bool("fAlSE"), Ok(false));
-        assert_eq!(parse_bool("f"), Ok(false));
-        assert_eq!(parse_bool("F"), Ok(false));
-        assert_eq!(parse_bool("no"), Ok(false));
-        assert_eq!(parse_bool("NO"), Ok(false));
-        assert_eq!(parse_bool("n"), Ok(false));
-        assert_eq!(parse_bool("N"), Ok(false));
-        assert_eq!(parse_bool("0"), Ok(false));
-        assert_eq!(parse_bool("000"), Ok(false));
-    }
-
-    #[test]
-    fn parse_bool_errors() {
-        assert!(parse_bool("X").is_err());
-        assert!(parse_bool("yes or no").is_err());
-        assert!(parse_bool("123.4").is_err());
-    }
 }
